+++ js/embedded/main.js
@@ -5,12 +5,12 @@
  * later. See the COPYING file.
  *
  * @author Pauli Järvinen <pauli.jarvinen@gmail.com>
- * @copyright Pauli Järvinen 2017 - 2023
+ * @copyright Pauli Järvinen 2017 - 2024
  */
 
 import playIconPath from '../../img/play-big.svg';
+import playIconSvgData from '../../img/play-big.svg?raw';
 import playOverlayPath from '../../img/play-overlay.svg';
-
 
 window.addEventListener('DOMContentLoaded', function() {
 	// Nextcloud 13+ have a built-in Music player in its "individual shared music file" page.
@@ -24,7 +24,7 @@ function initEmbeddedPlayer() {
 
 	let mCurrentFile = null; // may be an audio file or a playlist file
 	let mPlayingListFile = false;
-	let mFileList = null; // FileList from Files or Sharing app
+	let mFileList = null; // FileList from Files (prior to NC28) or Sharing app
 	let mShareToken = $('#sharingToken').val(); // undefined when not on share page
 
 	let mPlayer = new OCA.Music.EmbeddedPlayer(onClose, onNext, onPrev, onMenuOpen, onShowList, onImportList, onImportRadio);
@@ -52,9 +52,12 @@ function initEmbeddedPlayer() {
 	register();
 
 	function urlForFile(file) {
-		let url = mFileList.getDownloadUrl(file.name, file.path);
+		let url = mFileList
+			? mFileList.getDownloadUrl(file.name, file.path)
+			: OC.filePath('music', '', 'index.php') + '/api/file/' + file.id + '/download';
 
-		// append request token unless this is a public share
+		// Append request token unless this is a public share. This is actually unnecessary for most files
+		// but needed when the file in question is played using our Aurora.js fallback player.
 		if (!mShareToken) {
 			let delimiter = _.includes(url, '?') ? '&' : '?';
 			url += delimiter + 'requesttoken=' + encodeURIComponent(OC.requestToken);
@@ -65,28 +68,32 @@ function initEmbeddedPlayer() {
 	function onClose() {
 		mCurrentFile = null;
 		mPlayingListFile = false;
-		mPlaylist.reset();
-		if (OCA.Music.playlistTabView) {
-			OCA.Music.playlistTabView.setCurrentTrack(null, null);
-		}
+		mPlaylist?.reset();
+		OCA.Music.playlistTabView?.setCurrentTrack(null, null);
 	}
 
 	function onNext() {
-		jumpToPlaylistFile(mPlaylist.next());
+		if (mPlaylist) {
+			jumpToPlaylistFile(mPlaylist.next());
+		}
 	}
 
 	function onPrev() {
-		jumpToPlaylistFile(mPlaylist.prev());
+		if (mPlaylist) {
+			jumpToPlaylistFile(mPlaylist.prev());
+		}
 	}
 
 	function viewingCurrentFileFolder() {
+		// Note: this is always false on NC28+
 		return mCurrentFile && mFileList && mCurrentFile.path == mFileList.breadcrumb.dir;
 	}
 
 	function onMenuOpen($menu) {
 		// disable/enable the "Show list" item
 		let $showItem = $menu.find('#playlist-menu-show');
-		if (viewingCurrentFileFolder()) {
+		// the new sidebar API introduced in NC18 enabled viewing details also for files outside the current folder
+		if (OCA.Files.Sidebar || viewingCurrentFileFolder()) {
 			$showItem.removeClass('disabled');
 			$showItem.removeAttr('title');
 		} else {
@@ -96,7 +103,7 @@ function initEmbeddedPlayer() {
 
 		// disable/enable the "Import list to Music" item
 		let inLibraryFilesCount = _(mPlaylist.files()).filter('in_library').size();
-		let extStreamsCount = _(mPlaylist.files()).filter('url').size();
+		let extStreamsCount = _(mPlaylist.files()).filter('external').size();
 		let outLibraryFilesCount = mPlaylist.length() - inLibraryFilesCount;
 
 		let $importListItem = $menu.find('#playlist-menu-import');
@@ -132,8 +139,16 @@ function initEmbeddedPlayer() {
 	}
 
 	function onShowList() {
-		mFileList.scrollTo(mCurrentFile.name);
-		mFileList.showDetailsView(mCurrentFile.name, OCA.Music.playlistTabView.id);
+		if (OCA.Files.Sidebar) {
+			// This API is available starting from NC18 and after NC28, it's the only one available.
+			// This is better than the older API because this can be used also for files which are not
+			// present in the currently viewed folder.
+			OCA.Files.Sidebar.open(mCurrentFile.path + '/' + mCurrentFile.name);
+			OCA.Files.Sidebar.setActiveTab(OCA.Music.playlistTabView.id);
+		} else {
+			mFileList.scrollTo(mCurrentFile.name);
+			mFileList.showDetailsView(mCurrentFile.name, OCA.Music.playlistTabView.id);
+		}
 	}
 
 	function onImportList() {
@@ -145,18 +160,10 @@ function initEmbeddedPlayer() {
 	}
 
 	function doImportFromFile(serviceImportFunc) {
-		// The busy animation is shown on the file item if we are still viewing the folder
-		// where the file resides. The importing itself is possible regardless.
-		let $file = null;
-		if (viewingCurrentFileFolder()) {
-			$file = mFileList.findFileEl(mCurrentFile.name);
-			mFileList.showFileBusyState($file, true);
-		}
+		mPlayer.showBusy(true);
 
-		serviceImportFunc(mCurrentFile, function(_result) {
-			if ($file) {
-				mFileList.showFileBusyState($file, false);
-			}
+		serviceImportFunc(mCurrentFile, (_result) => {
+			mPlayer.showBusy(false);
 		});
 	}
 
@@ -167,7 +174,7 @@ function initEmbeddedPlayer() {
 			if (!mPlayingListFile) {
 				mCurrentFile = file;
 			}
-			if ('url' in file) {
+			if (file.external) {
 				mPlayer.playExtUrl(file.url, file.caption, mShareToken);
 			} else {
 				mPlayer.playFile(
@@ -192,8 +199,8 @@ function initEmbeddedPlayer() {
 		if (typeof OCA.Files !== 'undefined') {
 			OCA.Music.initPlaylistTabView(mPlaylistMimes);
 			connectPlaylistTabViewEvents();
-			registerFolderPlayer(mAudioMimes, openAudioFile);
-			registerFolderPlayer(mPlaylistMimes, openPlaylistFile);
+			registerFolderPlayer(mAudioMimes, openAudioFile, 'music_play_audio_file');
+			registerFolderPlayer(mPlaylistMimes, openPlaylistFile, 'music_play_playlist_file');
 		}
 
 		// Add player on single-file-share page if the MIME is a supported audio type
@@ -204,8 +211,8 @@ function initEmbeddedPlayer() {
 
 	function connectPlaylistTabViewEvents() {
 		if (OCA.Music.playlistTabView) {
-			OCA.Music.playlistTabView.on('playlistItemClick', function(playlistId, playlistName, itemIdx) {
-				if (mCurrentFile !== null && playlistId == mCurrentFile.id) {
+			OCA.Music.playlistTabView.on('playlistItemClick', (playlistFile, itemIdx) => {
+				if (mCurrentFile !== null && playlistFile.id == mCurrentFile.id) {
 					if (itemIdx == mPlaylist.currentIndex()) {
 						mPlayer.togglePlayback();
 					} else {
@@ -213,14 +220,18 @@ function initEmbeddedPlayer() {
 					}
 				}
 				else {
-					mFileList = OCA.Files.App.fileList;
-					mCurrentFile = mFileList.findFile(playlistName);
-					openPlaylistFile(function() {
-						jumpToPlaylistFile(mPlaylist.jumpToIndex(itemIdx));
-					});
+					if (OCA.Files.App) {
+						// Before NC28
+						mFileList = OCA.Files.App.fileList;
+						mCurrentFile = mFileList.findFile(playlistFile.name);
+					} else {
+						// NC28 or later
+						mCurrentFile = playlistFile;
+					}
+					openPlaylistFile(() => jumpToPlaylistFile(mPlaylist.jumpToIndex(itemIdx)));
 				}
 			});
-			OCA.Music.playlistTabView.on('rendered', function() {
+			OCA.Music.playlistTabView.on('rendered', () => {
 				if (mCurrentFile !== null) {
 					OCA.Music.playlistTabView.setCurrentTrack(mCurrentFile.id, mPlaylist.currentIndex());
 				}
@@ -231,7 +242,78 @@ function initEmbeddedPlayer() {
 	/**
 	 * "Folder player" is used in the Files app and on shared folders to play audio files and playlist files
 	 */
-	function registerFolderPlayer(mimes, openFileCallback) {
+	function registerFolderPlayer(mimes, openFileCallback, actionId) {
+		const wrappedCallback = function(file) {
+			// Check if playing file changes
+			if (mCurrentFile === null || mCurrentFile.id != file.id) {
+				mCurrentFile = file;
+				openFileCallback();
+			}
+			else {
+				mPlayer.togglePlayback();
+			}
+		};
+
+		if (OCA.Files?.fileActions) {
+			registerFolderPlayerBeforeNC28(mimes, wrappedCallback, actionId);
+		} else {
+			registerFolderPlayerAfterNC28(mimes, wrappedCallback, actionId);
+		}
+	}
+
+	function registerFolderPlayerAfterNC28(mimes, onActionCallback, actionId) {
+		import('@nextcloud/files').then(ncFiles => {
+			ncFiles.registerFileAction(new ncFiles.FileAction({
+				id: actionId,
+				displayName: () => t('music', 'Play'),
+				iconSvgInline: () => playIconSvgData,
+				default: ncFiles.DefaultType.DEFAULT,
+				order: -1, // prioritize over the built-in Viewer app
+
+				enabled: (nodes, _view) => {
+					if (nodes.length !== 1) {
+						return false;
+					}	
+					return mimes.includes(nodes[0].mime);
+				},
+
+				/**
+				 * Function executed on single file action
+				 * @return true if the action was executed successfully,
+				 * false otherwise and null if the action is silent/undefined.
+				 * @throws Error if the action failed
+				 */
+				exec: (file, view, dir) => {
+					const adaptFile = (f) => {
+						return {id: f.fileid, name: f.basename, mimetype: f.mime, path: dir};
+					};
+					onActionCallback(adaptFile(file));
+
+					if (!mPlayingListFile) {
+						// get the directory contents and use them as the play queue
+						view.getContents(dir).then(contents => {
+							const dirFiles = _.map(contents.contents, adaptFile);
+							// By default, the files are sorted simply by the character codes, putting upper case names before lower case
+							// and not respecting any locale settings. This doesn't match the order on the UI, regardless of the column
+							// used for sorting. Sort on our own treating numbers "naturally" and using the locale of the browser since
+							// this is how NC28 seems to do this (older NC versions, on the other hand, used the user-selected UI-language
+							// as the locale for sorting although user-selected locale would have made even more sense).
+							// This still leaves such a mismatch that the special characters may be sorted differently by localeCompare than
+							// what NC28 Files does (it uses the 3rd party library natural-orderby for this).
+							dirFiles.sort((a, b) => a.name.localeCompare(b.name, undefined, {numeric: true, sensitivity: 'base'}));
+	
+							mPlaylist = new OCA.Music.Playlist(dirFiles, mAudioMimes, mCurrentFile.id);
+							mPlayer.setNextAndPrevEnabled(mPlaylist.length() > 1);
+						});
+					}
+
+					return true;
+				},
+			}));
+		});
+	}
+
+	function registerFolderPlayerBeforeNC28(mimes, onActionCallback, actionId) {
 		// Handle 'play' action on file row
 		let onPlay = function(fileName, context) {
 			mFileList = context.fileList;
@@ -241,27 +323,20 @@ function initEmbeddedPlayer() {
 			// the user navigates to an audio file with a direct link. In that case, the callback happens before
 			// the context.filList is populated and we can't operate normally. Just ignore these cases.
 			if (file !== null) {
-				// Check if playing file changes
-				if (mCurrentFile === null || mCurrentFile.id != file.id) {
-					mCurrentFile = file;
-					openFileCallback();
-				}
-				else {
-					mPlayer.togglePlayback();
-				}
+				onActionCallback(file);
 			}
 		};
 
 		let registerPlayerForMime = function(mime) {
 			OCA.Files.fileActions.register(
 					mime,
-					'music-play',
+					actionId,
 					OC.PERMISSION_READ,
 					playIconPath,
 					onPlay,
 					t('music', 'Play')
 			);
-			OCA.Files.fileActions.setDefault(mime, 'music-play');
+			OCA.Files.fileActions.setDefault(mime, actionId);
 		};
 		_.forEach(mimes, registerPlayerForMime);
 	}
@@ -270,50 +345,63 @@ function initEmbeddedPlayer() {
 		mPlayingListFile = false;
 
 		mPlayer.show();
-		mPlaylist = new OCA.Music.Playlist(mFileList.files, mAudioMimes, mCurrentFile.id);
+		mPlayer.showBusy(false);
+		mPlaylist = new OCA.Music.Playlist(mFileList?.files ?? [mCurrentFile], mAudioMimes, mCurrentFile.id);
 		mPlayer.setNextAndPrevEnabled(mPlaylist.length() > 1);
 		jumpToPlaylistFile(mPlaylist.currentFile());
 	}
 
 	function openPlaylistFile(onReadyCallback = null) {
 		mPlayingListFile = true;
-		let $file = mFileList.findFileEl(mCurrentFile.name);
 
-		mFileList.showFileBusyState($file, true);
-		let onPlaylistLoaded = function(data) {
-			if (data.files.length > 0) {
-				mPlayer.show(mCurrentFile.name);
-				mPlaylist = new OCA.Music.Playlist(data.files, mAudioMimes, data.files[0].id);
-				mPlayer.setNextAndPrevEnabled(mPlaylist.length() > 1);
-				jumpToPlaylistFile(mPlaylist.currentFile());
-			}
-			else {
-				mCurrentFile = null;
-				OC.Notification.showTemporary(t('music', 'No files from the playlist could be found'));
-			}
-			if (data.invalid_paths.length > 0) {
-				let note = t('music', 'The playlist contained {count} invalid path(s).',
-						{count: data.invalid_paths.length});
-				if (!mShareToken) {
-					// Guide the user to look for details, unless this is a public share where the
-					// details pane is not available.
-					note += ' ' +  t('music', 'See the playlist file details.');
+		// clear the previous playback
+		mPlayer.stop();
+		mPlaylist = null;
+
+		mPlayer.show(mCurrentFile.name);
+		mPlayer.showBusy(true);
+
+		const listFileId = mCurrentFile.id;
+		const onPlaylistLoaded = function(data) {
+			// ignore the callback if the player is already closed or file changed by the time we get it
+			if (mCurrentFile?.id == listFileId) {
+				mPlayer.showBusy(false);
+				if (data.files.length > 0) {
+					mPlaylist = new OCA.Music.Playlist(data.files, mAudioMimes, data.files[0].id);
+					mPlayer.setNextAndPrevEnabled(mPlaylist.length() > 1);
+					jumpToPlaylistFile(mPlaylist.currentFile());
 				}
-				OC.Notification.showTemporary(note);
-			}
+				else {
+					mCurrentFile = null;
+					mPlayer.close();
+					OC.Notification.showTemporary(t('music', 'No files from the playlist could be found'));
+				}
+				if (data.invalid_paths.length > 0) {
+					let note = t('music', 'The playlist contained {count} invalid path(s).',
+							{count: data.invalid_paths.length});
+					if (!mShareToken) {
+						// Guide the user to look for details, unless this is a public share where the
+						// details pane is not available.
+						note += ' ' +  t('music', 'See the playlist file details.');
+					}
+					OC.Notification.showTemporary(note);
+				}
 
-			mFileList.showFileBusyState($file, false);
-
-			if (onReadyCallback) {
-				onReadyCallback();
+				if (onReadyCallback) {
+					onReadyCallback();
+				}
 			}
 		};
-		let onError = function() {
-			mCurrentFile = null;
-			OC.Notification.showTemporary(t('music', 'Error reading playlist file'));
-			mFileList.showFileBusyState($file, false);
+		const onError = function() {
+			// ignore the callback if the player is already closed or file changed by the time we get it
+			if (mCurrentFile?.id == listFileId) {
+				mPlayer.close();
+				mPlayer.showBusy(false);
+				mCurrentFile = null;
+				OC.Notification.showTemporary(t('music', 'Error reading playlist file'));
+			}
 		};
-		OCA.Music.PlaylistFileService.readFile(mCurrentFile.id, onPlaylistLoaded, onError, mShareToken);
+		OCA.Music.PlaylistFileService.readFile(listFileId, onPlaylistLoaded, onError, mShareToken);
 	}
 
 	/**