+++ lib/Controller/PlaylistApiController.php
@@ -355,6 +355,7 @@ class PlaylistApiController extends Controller {
 				if (isset($fileInfo['url'])) {
 					$fileInfo['id'] = $bogusUrlId--;
 					$fileInfo['mimetype'] = null;
+					$fileInfo['external'] = true;
 					return $fileInfo;
 				} else {
 					$file = $fileInfo['file'];
@@ -364,7 +365,8 @@ class PlaylistApiController extends Controller {
 						'path' => $this->userFolder->getRelativePath($file->getParent()->getPath()),
 						'mimetype' => $file->getMimeType(),
 						'caption' => $fileInfo['caption'],
-						'in_library' => isset($libFileIds[$file->getId()])
+						'in_library' => isset($libFileIds[$file->getId()]),
+						'external' => false
 					];
 				}
 			}, $result['files']);