+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [Unreleased]
 
 ### Added
+- Support for Nextcloud 28
+  [#1116](https://github.com/owncloud/music/pull/1116)
 - Ampache API: 
   * Support for argument `random` in the method `playlist_songs`
   * Method `bookmark`
@@ -22,6 +24,7 @@
 - Updated the getID3 library to the development version 1.9.23-202312292105
   * Fixes the issue of garbage bytes being extracted from some RIFF tags
     [#1115](https://github.com/owncloud/music/issues/1115)
+- Search within the Music app now works with an own input field in the navigation pane instead of the unified search input
 
 ### Fixed
 - Songs with scanned integer property value (like track number) larger than 2147483647 causing error on PostgreSQL